fix(serve-dir): escape XSS in HtmlFileList (GHSA-cwv4-h3j5-w3cf) gate listing behind `html` feature; render via rama_http::html
