+++ rama-http/src/service/fs/serve_dir/open_file.rs
@@ -5,28 +5,35 @@ use super::{
 use crate::headers::{encoding::Encoding, specifier::QualityValue};
 use crate::{HeaderValue, Method, Request, Uri, header};
 use http_range_header::RangeUnsatisfiableError;
-use jiff::Zoned;
 use rama_core::combinators::Either;
 use rama_core::telemetry::tracing;
 use rama_http_types::mime::Mime;
-use rama_utils::include_dir::{self, Dir, Metadata as EmbeddedMetadata};
+use rama_utils::include_dir::{Dir, Metadata as EmbeddedMetadata};
 use std::io::Cursor;
 use std::{
     ffi::OsStr,
-    fmt,
     fs::Metadata,
     io::{self, ErrorKind, SeekFrom},
     ops::RangeInclusive,
     path::{Path, PathBuf},
-    time::SystemTime,
 };
 use tokio::io::AsyncRead;
 use tokio::{fs::File, io::AsyncSeekExt};
 
+// All html-only state (DirEntry, HumanSize, generate_directory_html, the
+// emoji/size helpers, the per-helper unit tests) lives in a sibling
+// file pulled in via `#[path]` so the gating is paid for exactly once.
+#[cfg(feature = "html")]
+#[path = "open_file_html.rs"]
+mod html;
+
 /// Represents the outcome of attempting to open a file for serving.
 pub(super) enum OpenFileOutput {
     FileOpened(Box<FileOpened>),
-    Redirect { location: HeaderValue },
+    Redirect {
+        location: HeaderValue,
+    },
+    #[cfg(feature = "html")]
     Html(String),
     FileNotFound,
     PreconditionFailed,
@@ -521,127 +528,10 @@ async fn maybe_serve_directory(
             }
         }
         DirectoryServeMode::NotFound => Ok(Some(OpenFileOutput::FileNotFound)),
+        #[cfg(feature = "html")]
         DirectoryServeMode::HtmlFileList => {
-            let mut entries = vec![];
-
-            match source {
-                DirSource::Filesystem(_) => {
-                    let mut dir = tokio::fs::read_dir(&path_to_file).await?;
-                    while let Some(entry) = dir.next_entry().await? {
-                        let file_name = entry.file_name();
-                        let file_name_str = file_name.to_string_lossy().to_string();
-
-                        let metadata = entry.metadata().await?;
-                        let is_dir = metadata.is_dir();
-                        let modified = metadata.modified().unwrap_or(SystemTime::UNIX_EPOCH);
-                        let size = if is_dir { 0 } else { metadata.len() };
-
-                        entries.push(DirEntry::new(file_name_str, is_dir, modified, size));
-                    }
-                }
-                DirSource::Embedded(base) => {
-                    let Some(dir) = base.get_dir(path_to_file) else {
-                        return Ok(Some(OpenFileOutput::FileNotFound));
-                    };
-
-                    // Process all entries (directories and files)
-                    for entry in dir.entries() {
-                        let file_name_str = entry
-                            .path()
-                            .file_name()
-                            .map(|n| n.to_string_lossy().to_string())
-                            .unwrap_or_default();
-
-                        match entry {
-                            include_dir::DirEntry::Dir(_) => {
-                                let modified = SystemTime::UNIX_EPOCH;
-                                entries.push(DirEntry::new(file_name_str, true, modified, 0));
-                            }
-                            include_dir::DirEntry::File(file) => {
-                                let modified = file
-                                    .metadata()
-                                    .map(|m| m.modified())
-                                    .unwrap_or(SystemTime::UNIX_EPOCH);
-
-                                entries.push(DirEntry::new(
-                                    file_name_str,
-                                    false,
-                                    modified,
-                                    file.contents().len() as u64,
-                                ));
-                            }
-                        }
-                    }
-                }
-            }
-
-            let html = generate_directory_html(entries, uri);
-            Ok(Some(OpenFileOutput::Html(html)))
+            html::serve_html_listing(path_to_file, uri, source).await
         }
-    }
-}
-
-/// Human-readable file size representation.
-enum HumanSize {
-    None,
-    Bytes(u64),
-    KiloBytes(f64),
-    MegaBytes(f64),
-    GigaBytes(f64),
-    TeraBytes(f64),
-}
-
-impl fmt::Display for HumanSize {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        match self {
-            Self::None => write!(f, "--"),
-            Self::Bytes(n) => write!(f, "{n}B"),
-            Self::KiloBytes(d) => write!(f, "{d:.1}KB"),
-            Self::MegaBytes(d) => write!(f, "{d:.1}MB"),
-            Self::GigaBytes(d) => write!(f, "{d:.1}GB"),
-            Self::TeraBytes(d) => write!(f, "{d:.1}TB"),
-        }
-    }
-}
-
-/// Format file size in human-readable units (B, KB, MB, GB, TB).
-fn format_size(bytes: u64) -> HumanSize {
-    const MAX_UNITS: usize = 5;
-
-    let mut size = bytes as f64;
-    let mut unit = 0;
-
-    while size >= 1024.0 && unit < MAX_UNITS - 1 {
-        size /= 1024.0;
-        unit += 1;
-    }
-    match unit {
-        0 => HumanSize::Bytes(size as u64),
-        1 => HumanSize::KiloBytes(size),
-        2 => HumanSize::MegaBytes(size),
-        3 => HumanSize::GigaBytes(size),
-        _ => HumanSize::TeraBytes(size),
-    }
-}
-
-/// Get an appropriate emoji icon for a file based on its MIME type.
-fn emoji_for_mime(mime: Option<&crate::mime::Mime>, is_dir: bool) -> &'static str {
-    if is_dir {
-        return "📁";
-    }
-
-    match mime.map(|m| (m.type_().as_str(), m.subtype().as_str())) {
-        Some(("text", "css")) => "🎨",
-        Some(("image", _)) => "🖼️",
-        Some(("audio", _)) => "🎵",
-        Some(("video", _)) => "🎬",
-        Some(("application", "pdf")) => "📕",
-        Some(("application", "zip" | "x-tar")) => "🗜️",
-        Some(("application", "json" | "xml")) => "🔧",
-        Some(("application", "msword")) => "📃",
-        Some(("application", "vnd.ms-excel")) => "📊",
-        Some(("application", "javascript")) => "🧩",
-        _ => "📄",
     }
 }
 
@@ -714,226 +604,4 @@ fn append_slash_on_path(uri: Uri) -> Result<Uri, OpenFileOutput> {
         tracing::error!("redirect uri failed to build: {err:?}");
         OpenFileOutput::InvalidRedirectUri
     })
-}
-
-/// Represents a directory entry for HTML file listing.
-struct DirEntry {
-    name: String,
-    is_dir: bool,
-    modified: SystemTime,
-    size: u64,
-}
-
-impl DirEntry {
-    fn new(name: String, is_dir: bool, modified: SystemTime, size: u64) -> Self {
-        Self {
-            name,
-            is_dir,
-            modified,
-            size,
-        }
-    }
-}
-
-/// Generate HTML page for directory listing with file details and navigation.
-fn generate_directory_html(entries: Vec<DirEntry>, uri: &Uri) -> String {
-    let mut rows = vec![];
-
-    for entry in entries {
-        let modified_str = format_system_time_local(entry.modified);
-
-        let mime = if entry.is_dir {
-            None
-        } else {
-            crate::mime::guess::from_path(entry.name.as_str()).first()
-        };
-        let emoji = emoji_for_mime(mime.as_ref(), entry.is_dir);
-
-        let hs = if entry.is_dir {
-            HumanSize::None
-        } else {
-            format_size(entry.size)
-        };
-
-        rows.push(format!(
-            "<tr><td>{5} <a href=\"{1}{2}{0}\">{0}</a></td><td>{3}</td><td>{4}</td></tr>",
-            entry.name,
-            uri.path().trim_end_matches('/'),
-            if uri.path().trim_start_matches('/').is_empty() {
-                ""
-            } else {
-                "/"
-            },
-            modified_str,
-            hs,
-            emoji,
-        ));
-    }
-
-    let table = format!(
-        r#"<table style="width:100%; border-collapse:collapse;">
-        <thead>
-        <tr><th align="left">Name</th><th align="left">Last Modified</th><th align="left">Size</th></tr>
-        </thead>
-        <tbody>
-        {0}
-        </tbody>
-        </table>"#,
-        rows.join("\n")
-    );
-
-    let mut nav_parts = vec![];
-    let mut current_path = String::new();
-    for part in uri.path().trim_start_matches('/').split('/') {
-        if !part.is_empty() {
-            current_path.push('/');
-            current_path.push_str(part);
-            nav_parts.push(format!("<a href=\"{current_path}\">{part}</a>"));
-        }
-    }
-    let breadcrumb = if nav_parts.is_empty() {
-        "<a href=\"/\">/</a>".to_owned()
-    } else {
-        format!(
-            "<a href=\"/\">/</a> &raquo; {}",
-            nav_parts.join(" &raquo; ")
-        )
-    };
-
-    format!(
-        r#"<!DOCTYPE HTML>
-        <html lang="en">
-        <head>
-        <meta charset="utf-8">
-        <title>Directory listing for .{0}</title>
-        </head>
-        <body>
-        <h1>Directory listing for .{0}</h1>
-        <div>{2}</div>
-        <hr>
-        <ul>
-        {1}
-        </ul>
-        <hr>
-        </body>
-        </html>"#,
-        uri.path(),
-        table,
-        breadcrumb,
-    )
-}
-
-fn format_system_time_local(system_time: SystemTime) -> String {
-    Zoned::try_from(system_time)
-        .map(|zdt| zdt.strftime("%Y-%m-%d %H:%M:%S %:z").to_string())
-        .unwrap_or_else(|_| "-".to_owned())
-}
-
-#[cfg(test)]
-mod test {
-    use std::str::FromStr;
-
-    use super::*;
-    use crate::mime::Mime;
-
-    #[test]
-    fn test_emoji_for_mime() {
-        struct Case {
-            mime: Option<Mime>,
-            is_dir: bool,
-            expected: &'static str,
-        }
-
-        let cases = [
-            Case {
-                mime: None,
-                is_dir: true,
-                expected: "📁",
-            },
-            Case {
-                mime: Some(Mime::from_str("text/plain").unwrap()),
-                is_dir: false,
-                expected: "📄",
-            },
-            Case {
-                mime: Some(Mime::from_str("image/png").unwrap()),
-                is_dir: false,
-                expected: "🖼️",
-            },
-            Case {
-                mime: Some(Mime::from_str("audio/mpeg").unwrap()),
-                is_dir: false,
-                expected: "🎵",
-            },
-            Case {
-                mime: Some(Mime::from_str("application/pdf").unwrap()),
-                is_dir: false,
-                expected: "📕",
-            },
-            Case {
-                mime: Some(Mime::from_str("application/zip").unwrap()),
-                is_dir: false,
-                expected: "🗜️",
-            },
-            Case {
-                mime: Some(Mime::from_str("application/json").unwrap()),
-                is_dir: false,
-                expected: "🔧",
-            },
-            Case {
-                mime: Some(Mime::from_str("application/octet-stream").unwrap()),
-                is_dir: false,
-                expected: "📄",
-            },
-        ];
-
-        for case in cases {
-            let actual = emoji_for_mime(case.mime.as_ref(), case.is_dir);
-            assert_eq!(actual, case.expected, "Failed on case: {:?}", case.mime);
-        }
-    }
-
-    #[test]
-    fn test_format_size() {
-        struct Case {
-            input: u64,
-            expected: &'static str,
-        }
-
-        let cases = [
-            Case {
-                input: 0,
-                expected: "0B",
-            },
-            Case {
-                input: 512,
-                expected: "512B",
-            },
-            Case {
-                input: 1023,
-                expected: "1023B",
-            },
-            Case {
-                input: 1024,
-                expected: "1.0KB",
-            },
-            Case {
-                input: 1048576,
-                expected: "1.0MB",
-            },
-            Case {
-                input: 1073741824,
-                expected: "1.0GB",
-            },
-            Case {
-                input: 1099511627776,
-                expected: "1.0TB",
-            },
-        ];
-
-        for case in cases {
-            let actual = format_size(case.input).to_string();
-            assert_eq!(actual, case.expected, "Failed on input: {}", case.input);
-        }
-    }
 }
